fpl-ml-system/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Caching==2.1.0

# Database
SQLAlchemy==2.0.23

# Data processing and ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
xgboost==2.0.3

# Optimization
PuLP==2.7.0

# HTTP and API
requests==2.31.0
httpx==0.25.0

# Validation
pydantic==2.5.2

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
ruff==0.1.8
black==23.11.0
mypy==1.7.1
isort==5.12.0

# Production server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0